// src/app/models/user.model.ts
export interface User {
  id: number;
  email: string;
  fullname: string;
  firstname: string;
  lastname: string;
  residenceCountry: User.Country;
  roles: {
    id: number;
    roleName: string;
  };

}
export interface UserForm {
  email: string;
  fullname: string;
  firstname: string;
  lastname: string;
  residenceCountry: User.Country;
  roles: { id: number; };
  familyLink: {
    id: number
  },
  currentFidelityRank: {
    id: number
  }
}